+++ src/test/java/conflict/guru/qa/secondConflictTest.java
@@ -6,13 +6,10 @@ import org.junit.jupiter.api.Test;
 
 import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Selenide.$;
-import static com.codeborne.selenide.Selenide.open;
-public class conflictTest {
+
+public class secondConflictTest {
     @Test
     void conflictTest(){
-        Configuration.pageLoadStrategy = "eager";
-        Selenide.open("https://google.ru");
-        $("[name=q]").setValue("Selenide").pressEnter();
-        $("#search").shouldHave(text("https://ru.selenide.org"));
+        Selenide.open("https://ya.ru");
     }
 }
